pyproject.toml
[tool.poetry]
name = "crypto-wallet"
version = "0.1.0"
description = ""
authors = ["dima_nikitin <nikitindima1206@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.89.1"
greenlet = "^2.0.1"
sqlalchemy= "^1.4.46"
python-dotenv = "^0.21.0"
yarl = "^1.8.2"
sqlalchemy-utils = "^0.39.0"
alembic = "^1.9.1"
psycopg2-binary = "^2.9.5"
uvicorn = "^0.20.0"
asyncpg = "^0.27.0"
databases = "^0.7.0"
python-socketio = "^5.7.2"
async-lru = "^1.0.3"
email-validator = "^1.3.0"
fastapi-helper = "^0.0.12"
passlib = "^1.7.4"
PyJWT = "^2.6.0"
aio-pika = "^8.3.0"
toml = "^0.10.2"
Jinja2 = "^3.1.2"
beanie = "^1.16.8"
bcrypt = "^4.0.1"
pyhumps = "^3.8.0"
python-multipart = "^0.0.5"
pydantic = {extras = ["email"], version = "^1.10.4"}
aioredis = "^2.0.1"
flake8 = "^6.0.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
